Refactor : regardless to some classes & conventions rules
