+++ .gitignore
@@ -1,3 +1,5 @@
 public/assets/images
 src/admin/assets/images
 config/config.json
+.idea
+vendor
