+++ public/assets/css/blog.css
@@ -1,6 +1,7 @@
 ul {
   padding-left: 0;
 }
+
 header {
   padding-inline: 2em;
   padding-top: 2.5em;
@@ -22,8 +23,8 @@ header {
   top: 7px;
 }
 
-#blog{
-  margin-top:5em;
+#blog {
+  margin-top: 5em;
 }
 
 
@@ -32,6 +33,7 @@ header {
 #banner {
   width: 100%;
 }
+
 #banner img {
   width: 100%;
 
@@ -40,18 +42,18 @@ header {
 #list-article {
   display: flex;
   flex-direction: column;
-  gap:1em;
- width: 80%;
- 
+  gap: 1em;
+  width: 80%;
+
 }
 
-.article{
+.article {
   text-decoration: none;
-  color:var(--black);
+  color: var(--black);
 }
 
 .article img {
-  box-shadow: 5px 5px 9px rgba(0,0,0,.5);
+  box-shadow: 5px 5px 9px rgba(0, 0, 0, .5);
   border: .3em solid black;
 }
 
@@ -72,31 +74,31 @@ header {
   font-weight: var(--700);
 }
 
-.author-article{
+.author-article {
   text-align: center;
 }
 
-.row{
+.row {
   display: flex;
   flex-direction: column;
   align-items: center;
-  gap:5em;
+  gap: 5em;
 }
 
-.tags{
+.tags {
   display: flex;
   justify-content: center;
   padding-left: 0;
   gap: 1.5em;
 }
 
-.tags .tag-article{
+.tags .tag-article {
   font-weight: var(--700);
 }
 
 
 /* DESKTOP DISPLAY */
-.author-box{
+.author-box {
   display: flex;
   margin-bottom: 1em;
 }
@@ -106,28 +108,28 @@ header {
   border-radius: 50%;
 }
 
-.author-article{
-  align-self:center;
+.author-article {
+  align-self: center;
   position: relative;
-  top:5px;
-  left:.5em;
+  top: 5px;
+  left: .5em;
 }
 
 .article-short-phrase {
   color: #ada3a3;
 }
 
- .article-content {
-  text-align: left ;
+.article-content {
+  text-align: left;
 }
 
-.date-article{
+.date-article {
   font-weight: var(--700);
 }
 
-.dates-tags{
+.dates-tags {
   display: flex;
-  gap:1em;
+  gap: 1em;
 }
 
 
@@ -135,66 +137,68 @@ header {
 
 /* TABLET */
 @media (min-width:577px) and (max-width:991px) {
-  
-  .article-content{
+
+  .article-content {
     text-align: center;
   }
-  
+
   #list-article {
-    margin-top:6em;
+    margin-top: 6em;
   }
-  
+
 
   #banner img {
     height: 600px;
     transition: all 300ms ease;
   }
-  
-.article:visited{
-  color:var(--black);
-}
 
-.article img {
-  border-radius: 50%;
-  width: 200px;
-  margin: 0 auto;
-  display: block;
-  margin-bottom: 2em;
-}
+  .article:visited {
+    color: var(--black);
+  }
 
-.author-box{
-  flex-direction: column;
-}
+  .article img {
+    border-radius: 50%;
+    width: 200px;
+    margin: 0 auto;
+    display: block;
+    margin-bottom: 2em;
+  }
 
-.author-article{
-  left:0;
-}
+  .author-box {
+    flex-direction: column;
+  }
 
-.article-title{
-  text-align: center;
-  line-height: 2em;
-}
+  .author-article {
+    left: 0;
+  }
 
-.article-short-phrase {
-  text-align: center;
-  font-size: .8em;
-  color: #c0bfbf;
-}
-.article-content {
-  text-align: center;
-}
+  .article-title {
+    text-align: center;
+    line-height: 2em;
+  }
 
-.dates-tags{
-  display: flex;
-flex-direction: column;
-text-align: center;
-}
-.tags{
-  display: flex;
-  justify-content: center;
-  padding-left: 0;
-  gap: 1.5em;
-}
+  .article-short-phrase {
+    text-align: center;
+    font-size: .8em;
+    color: #c0bfbf;
+  }
+
+  .article-content {
+    text-align: center;
+  }
+
+  .dates-tags {
+    display: flex;
+    flex-direction: column;
+    text-align: center;
+  }
+
+  .tags {
+    display: flex;
+    justify-content: center;
+    padding-left: 0;
+    gap: 1.5em;
+  }
 
 
 
@@ -205,62 +209,62 @@ text-align: center;
 
   #banner img {
     height: 550px;
-   
+
   }
+
   #list-article {
-    margin-top:6em;
+    margin-top: 6em;
   }
-  
-  
-.article:visited{
-  color:var(--black);
-}
-
-.article img {
-border-radius: 50%;
-width: 200px;
-margin: 0 auto;
-display: block;
-margin-bottom: 2em;
-
-}
 
 
-.article-title{
-  text-align: center;
-  line-height: 2em;
-}
+  .article:visited {
+    color: var(--black);
+  }
 
-.article-box .article-short-phrase {
-  text-align: center;
-  font-size: .8em;
-  color: #c0bfbf;
-}
-.article-content {
-  text-align: center;
-}
+  .article img {
+    border-radius: 50%;
+    width: 200px;
+    margin: 0 auto;
+    display: block;
+    margin-bottom: 2em;
 
-.author-box{
-  flex-direction: column;
-}
-
-.author-article{
-  left:0;
-}
+  }
 
 
-.dates-tags{
-  display: flex;
-flex-direction: column;
-text-align: center;
-}
+  .article-title {
+    text-align: center;
+    line-height: 2em;
+  }
 
-.tags{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-left: 0;
-}
+  .article-box .article-short-phrase {
+    text-align: center;
+    font-size: .8em;
+    color: #c0bfbf;
+  }
+
+  .article-content {
+    text-align: center;
+  }
+
+  .author-box {
+    flex-direction: column;
+  }
+
+  .author-article {
+    left: 0;
+  }
 
 
+  .dates-tags {
+    display: flex;
+    flex-direction: column;
+    text-align: center;
+  }
+
+  .tags {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding-left: 0;
+  }
 }
