+++ public/assets/css/general.css
@@ -78,6 +78,5 @@ footer {
   footer p,
   footer a {
     font-size: .7em;
-
   }
 }
