+++ public/assets/css/homepage.css
@@ -149,10 +149,11 @@ header {
 /* DESKTOP */
 @media (min-width:992px) {
 
-  #starter-block .row{
-    padding-top:8em;
+  #starter-block .row {
+    padding-top: 8em;
     transition: all ease-in-out 300ms;
-  } 
+  }
+
   #form-box {
     margin: 0 auto;
   }
@@ -161,17 +162,18 @@ header {
     width: 400px;
   }
 
-  #about-text{
+  #about-text {
     width: 50%;
     margin: 0 auto;
     padding-bottom: 2em;
     transition: all 300ms ease-in-out;
   }
 
-    #title-inner-phrase{
-      padding-top: .7em;
-      font-size: .7em;
-    }
+  #title-inner-phrase {
+    padding-top: .7em;
+    font-size: .7em;
+  }
+
   #social-media-content {
     display: flex;
     flex-direction: column;
@@ -179,17 +181,18 @@ header {
     gap: 2.5em;
   }
 
-  #linkedin,#github{
+  #linkedin,
+  #github {
     width: 400px;
   }
 }
 
 /*TABLET */
 @media (min-width:320px) and (max-width:991px) {
-  #starter-block .row{
-    padding-top:2em;
+  #starter-block .row {
+    padding-top: 2em;
     transition: all ease-in-out 300ms;
-  } 
+  }
 
   #h1 {
     font-size: 1.8em;
@@ -198,7 +201,7 @@ header {
     transition: all 300ms ease-in-out;
   }
 
-  #title-inner-phrase{
+  #title-inner-phrase {
     /* padding-top: .7em; */
     font-size: .6em;
   }
