+++ src/Controller/ArticleController.php
@@ -2,13 +2,16 @@
 
 namespace Controller;
 
-use Model\ArticleModel;
+use Config\DatabaseConnection;
+use Model\Article;
+
 
 class ArticleController
 {
-  public function listOfAllArticles() : array
+  public function listOfAllArticles(): array
   {
-    $articles = new ArticleModel();
+    $db = new DatabaseConnection("professional_blog", "root", "");
+    $articles = new Article($db);
     return $articles->getArticles();
   }
 }
