+++ src/admin/assets/css/admin_add_article.css
@@ -13,12 +13,12 @@ header {
   font-size: 1.5em;
 }
 
-#article{
+#article {
   width: 80%;
   margin-top: 10em;
 }
 
-#banner{
+#banner {
   margin-bottom: 10em;
 }
 
@@ -26,11 +26,11 @@ header {
   width: 100%;
 }
 
-.requirements-box{
+.requirements-box {
   position: relative;
 }
 
-#h1{
+#h1 {
   border-bottom: 1px solid var(--orange);
   border-width: .2em;
   font-weight: var(--900);
@@ -39,35 +39,37 @@ header {
   padding-bottom: .4em;
 }
 
-.requirements{
+.requirements {
   position: absolute;
   right: 0;
   font-weight: var(--700);
   color: var(--black);
 }
-form{
+
+form {
   margin-top: 5em;
   margin-bottom: 5em;
 }
-.form-box{
-display: flex;
-flex-direction: column;
-gap:1.5em;
+
+.form-box {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5em;
 }
 
-.form-label{
+.form-label {
   font-weight: var(--700);
   color: var(--black);
 }
 
-.row .form-control{
+.row .form-control {
   border-color: var(--black);
-border-radius: 0;
+  border-radius: 0;
 }
 
-.row .form-control:focus{
+.row .form-control:focus {
   box-shadow: 0 0 0 .25rem var(--orange);
-border-color: rgba(0, 0, 0, .2);
+  border-color: rgba(0, 0, 0, .2);
 }
 
 .form-box .btn-dark {
@@ -75,49 +77,51 @@ border-color: rgba(0, 0, 0, .2);
   font-weight: var(--700);
 }
 
-#preview{
+#preview {
   margin-top: 10em;
   margin-bottom: 3em;
 }
 
-.row .preview-box{
+.row .preview-box {
   display: flex;
-flex-direction: column;
-gap: 1em;
-}
-#preview-title{
- 
-    border-bottom: 1px solid var(--orange);
-    border-width: .2em;
-    font-weight: var(--900);
-    color: var(--black);
-    text-align: center;
-    padding-bottom: .4em;
-    margin-bottom: 3em;
+  flex-direction: column;
+  gap: 1em;
 }
 
-#img-preview{
+#preview-title {
+
+  border-bottom: 1px solid var(--orange);
+  border-width: .2em;
+  font-weight: var(--900);
+  color: var(--black);
+  text-align: center;
+  padding-bottom: .4em;
+  margin-bottom: 3em;
+}
+
+#img-preview {
   width: 100%;
 }
 
-#title-preview{
-  margin-top:2em;
+#title-preview {
+  margin-top: 2em;
   text-align: center;
   font-weight: var(--900);
   padding-bottom: .5em;
 }
 
-#short-phrase-preview{
+#short-phrase-preview {
   color: #ada3a3;
-font-weight: var(--700);
-text-align: center;
+  font-weight: var(--700);
+  text-align: center;
 }
 
-#content-preview{
+#content-preview {
   line-height: 2.6em;
-font-size: 1.5em;
-font-weight: var(--700);
+  font-size: 1.5em;
+  font-weight: var(--700);
 }
+
 @media (min-width:577px) and (max-width:991px) {
 
   #content-preview {
