+++ src/admin/assets/css/admin_article_and_commentary.css
@@ -57,24 +57,26 @@ header {
   border-width: .2em;
   position: relative;
 }
-.article-title::after{
+
+.article-title::after {
   content: '';
   position: absolute;
   width: 100%;
   border-bottom: 1px solid orange;
-    border-bottom-width: 1px;
+  border-bottom-width: 1px;
   border-width: .2em;
   left: 50%;
   top: 50%;
-  transform: translate(-50%,-50%);
+  transform: translate(-50%, -50%);
   bottom: -24px;
 }
 
-.article-short-phrase{
+.article-short-phrase {
   text-align: center;
   color: #ada3a3;
 }
-.article-content{
+
+.article-content {
 
   line-height: 2em;
   margin-bottom: 2em;
@@ -101,14 +103,15 @@ header {
   margin-inline: 0;
 }
 
-#comment{
+#comment {
   margin-top: 5em;
   margin-bottom: 5em;
   width: 80%;
 }
-#comment-main-title{
+
+#comment-main-title {
   font-weight: var(--900);
-  color:var(--black);
+  color: var(--black);
   width: 100%;
   border-width: .2em;
   text-align: center;
@@ -117,39 +120,40 @@ header {
   position: relative;
 }
 
-#comment-main-title::after{
+#comment-main-title::after {
   content: '';
   position: absolute;
   width: 100%;
   border-bottom: 1px solid orange;
-    border-bottom-width: 1px;
+  border-bottom-width: 1px;
   border-width: .2em;
   left: 50%;
   top: 50%;
-  transform: translate(-50%,-50%);
+  transform: translate(-50%, -50%);
   bottom: -24px;
 }
 
-.commentary-box{
+.commentary-box {
   text-align: center;
 }
 
-.author-comment{
+.author-comment {
   font-weight: var(--700);
-  color:var(--black);
+  color: var(--black);
 }
 
-.content-comment{
+.content-comment {
   margin-bottom: 2.5em;
   text-align: justify;
 }
 
-form{
+form {
   display: flex;
   flex-direction: column;
-  gap:2em ;
+  gap: 2em;
 }
-form label{
+
+form label {
   font-weight: var(--700);
   padding-bottom: .8em;
 }
@@ -158,9 +162,10 @@ form textarea {
   resize: none;
 }
 
-#commands-article{
-  margin-top:2em;
+#commands-article {
+  margin-top: 2em;
 }
+
 .commands-admin {
   display: flex;
   justify-content: center;
@@ -168,10 +173,11 @@ form textarea {
 }
 
 
-.commands-admin .command{
+.commands-admin .command {
   font-weight: var(--700);
-  color:var(--white);
+  color: var(--white);
 }
+
 @media (max-width:576px) {
 
   #article .article-banner {
@@ -200,7 +206,7 @@ form textarea {
   }
 
   .article-date {
-   
+
     font-size: .9em;
   }
 }
