+++ src/admin/templates/admin_add_article.twig
@@ -98,4 +98,3 @@
 	</div>
 	<script src="../src/admin/assets/js/article_live_preview.js"></script>
 {% endblock %}
-
