+++ src/admin/templates/admin_article_and_commentary.twig
@@ -60,4 +60,3 @@
 		</form>
 	</div>
 {% endblock %}
-
