+++ src/admin/templates/admin_homepage.twig
@@ -69,7 +69,7 @@
 							<h4>CyranoB , le 19 juin 2023</h4>
 							<p>
 								C'est un roc !... c'est un pic... c'est un cap !
-																																Que dis-je, c'est un cap ?... C'est une péninsule ! »</p>
+																																								Que dis-je, c'est un cap ?... C'est une péninsule ! »</p>
 						</div>
 					</a>
 				</div>
@@ -133,4 +133,3 @@
 
 	</div>
 {% endblock %}
-
