+++ templates/article.twig
@@ -27,6 +27,7 @@
 						<img class="author-img" src="../public/assets/images/author_1.jpg" alt="avatar author1">
 						<h2 class="author-name">John Doe</h2>
 					</div>
+
 					<h1 class="article-title" id="h1">Les bases de PHP</h1>
 					<p class="article-short-phrase">Découvrez les fondamentaux de PHP pour développer des sites web dynamiques.</p>
 					<p class="article-content">PHP, acronyme de Hypertext Preprocessor, est un langage de script largement utilisé pour le développement web. Avec PHP, vous pouvez créer des pages web interactives et dynamiques en générant du contenu HTML en temps réel. Que vous souhaitiez créer un site web simple ou une application web complexe, les connaissances de base de PHP sont essentielles. Ce langage polyvalent offre de nombreuses fonctionnalités et facilite l'interaction avec les bases de données, la manipulation de fichiers, la gestion des formulaires et bien plus encore. Plongez dans l'univers de PHP et explorez ses possibilités infinies !</p>
@@ -89,4 +90,3 @@
 		</div>
 	</div>
 {% endblock %}
-
