+++ templates/blog.twig
@@ -68,4 +68,3 @@
 		</div>
 	</div>
 {% endblock %}
-
