+++ templates/homepage.twig
@@ -126,4 +126,3 @@
 
 	</div>
 </div>{% endblock %}
-
