+++ templates/layout.twig
@@ -19,4 +19,3 @@
 		{% include "footer.twig" %}
 	</body>
 </html>
-
