+++ templates/sign_in.twig
@@ -69,4 +69,3 @@
 	</div>
 
 {% endblock %}
-
