+++ templates/sign_up.twig
@@ -101,4 +101,3 @@
 	</div>
 	<script src="../public/assets/js/seePassword.js"></script>
 {% endblock %}
-
